Added and Edited some files

+++ app/src/main/java/com/example/cliwatchjc/data/tracker/PersonalGoalDetailsDao.kt
@@ -3,12 +3,14 @@ package com.example.cliwatchjc.data.tracker
 import androidx.room.Dao
 import androidx.room.Query
 import androidx.room.Update
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
 
 @Dao
 interface PersonalGoalDetailsDao {
 
     @Query("SELECT * FROM PersonalGoalDetails where userId = :userId AND goalId = :goalId")
-    suspend fun getGoalDetails(userId: Long, goalId: Long): List<PersonalGoalDetails>
+    suspend fun getGoalDetails(userId: Int, goalId: Long): List<PersonalGoalDetails>
 
     @Update
     suspend fun updateGoalDetails(goalDetails: PersonalGoalDetails)
